working on the project

+++ controllers/clothingItems.js
@@ -33,25 +33,38 @@ const createItem = (req, res) => {
 
 const deleteItem = (req, res) => {
   const { itemId } = req.params;
-  Item.findByIdAndDelete(itemId)
-    .orFail()
-    .then(() => res.send({ message: "item deleted succesesfully" }))
-    .catch((err) => {
-      console.error(err);
-      if (err.name === "DocumentNotFoundError") {
-        return res
-          .status(errorCode.notFound)
-          .send({ message: errorMessage.idNotFound });
-      }
-      if (err.name === "CastError") {
-        return res
-          .status(errorCode.badRequest)
-          .send({ message: errorMessage.badRequest });
-      }
+  const userId = req.user._id;
+  Item.findById(itemId).then((item) => {
+    if (!item) {
       return res
-        .status(errorCode.serverError)
-        .send({ message: errorMessage.defaultError });
-    });
+        .status(errorCode.notFound)
+        .send({ message: errorMessage.idNotFound });
+    }
+    if (item.owner.toString() !== userId) {
+      return res
+        .status(errorCode.forbidden)
+        .send({ message: "The item can`t be deleted" });
+    }
+    return Item.findByIdAndDelete(itemId)
+      .orFail()
+      .then(() => res.send({ message: "item deleted succesesfully" }))
+      .catch((err) => {
+        console.error(err);
+        if (err.name === "DocumentNotFoundError") {
+          return res
+            .status(errorCode.notFound)
+            .send({ message: errorMessage.idNotFound });
+        }
+        if (err.name === "CastError") {
+          return res
+            .status(errorCode.badRequest)
+            .send({ message: errorMessage.badRequest });
+        }
+        return res
+          .status(errorCode.serverError)
+          .send({ message: errorMessage.defaultError });
+      });
+  });
 };
 
 const likeItem = (req, res) => {
